pyproject.toml
[tool.poetry]
name = "market-project"
version = "0.1"
description = ""
authors = ["Георгий <nazimov03@list.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.43"
alembic = "^1.16.5"
uvicorn = "^0.35.0"
asyncpg = "^0.30.0"
pydantic-settings = "^2.10.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
authlib = "^1.6.3"
itsdangerous = "^2.2.0"
flask = "^3.1.2"
flask-admin = "^1.6.1"
pylint = "^3.3.8"
psycopg2-binary = "^2.9.10"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
gunicorn = "^23.0.0"
httpx = "^0.28.1"
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
freezegun = "^1.5.5"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
pylint = "^3.3.8"
locust = "^2.41.2"
isort = "<7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"
line_length = 88